+++ src/app/components/flows/queues/queueController.js
@@ -1,15 +1,13 @@
 'use strict';
 
 angular.module('liveopsConfigPanel')
-  .controller('QueueController', ['$rootScope', '$scope', 'Queue', 'Session', 'queueTableConfig', 'QueueVersion', 'Alert', 'Modal', '$translate', 'loEvents', 'ZermeloService',
-    function($rootScope, $scope, Queue, Session, queueTableConfig, QueueVersion, Alert, Modal, $translate, loEvents, ZermeloService) {
+  .controller('QueueController', ['$scope', 'Queue', 'Session', 'queueTableConfig', 'QueueVersion', 'Alert', '$translate', 'loEvents',
+    function($scope, Queue, Session, queueTableConfig, QueueVersion, Alert, $translate, loEvents) {
       var vm = this;
       vm.Session = Session;
       vm.tableConfig = queueTableConfig;
       vm.forms = {};
       vm.selectedQueueVersion = null;
-
-      vm.showBasicQuery = true;
 
       vm.fetchQueues = function() {
         return Queue.cachedQuery({
@@ -19,7 +17,7 @@ angular.module('liveopsConfigPanel')
 
       vm.getDefaultVersion = function() {
         return new QueueVersion({
-          query: ZermeloService.getQueryString(),
+          query: '[{:after-seconds-in-queue 0 :query {}}]',
           tenantId: Session.tenant.tenantId,
           name: 'v1',
           maxPriority: 1000,
@@ -33,8 +31,6 @@ angular.module('liveopsConfigPanel')
       $scope.$on(loEvents.tableControls.itemCreate, function() {
         vm.selectedQueueVersion = null;
 
-        $rootScope.$emit('queue.query.reset');
-
         vm.initialVersion = vm.getDefaultVersion();
 
         vm.selectedQueue = new Queue({
@@ -42,22 +38,9 @@ angular.module('liveopsConfigPanel')
         });
       });
 
-      $scope.$on(loEvents.bulkActions.close, vm.reset);
-
-      vm.confirmCancel = function() {
-        return Modal.showConfirm(
-          {
-            message: $translate.instant('unsavedchanges.nav.warning'),
-            okCallback: vm.reset
-          }
-        );
-      };
-
-      vm.reset = function() {
-        $rootScope.$emit('queue.query.reset');
+      $scope.$on(loEvents.bulkActions.close, function() {
         vm.selectedQueueVersion = null;
-        vm.selectedQueue = null;
-      };
+      });
 
       vm.submit = function() {
         var isNew = vm.selectedQueue.isNew();
@@ -77,11 +60,9 @@ angular.module('liveopsConfigPanel')
 
       vm.saveInitialVersion = function(queue) {
         var qv = vm.initialVersion;
-        qv.query = ZermeloService.getQueryString();
 
         vm.initialVersion.save()
           .then(function(versionResult) {
-            $rootScope.$emit('queue.query.reset');
             queue.activeVersion = versionResult.version;
             queue.activeQueue = versionResult;
             queue.active = true;
@@ -128,10 +109,8 @@ angular.module('liveopsConfigPanel')
       };
 
       vm.copySelectedVersion = function(version) {
-        $rootScope.$emit('queue.query.reset');
-        ZermeloService.parseString(version.query)
         vm.selectedQueueVersion = new QueueVersion({
-          query: ZermeloService.getQueryString(),
+          query: version.query,
           name: 'v' + (vm.fetchVersions().length + 1),
           tenantId: version.tenantId,
           queueId: version.queueId,
@@ -200,17 +179,14 @@ angular.module('liveopsConfigPanel')
       });
 
       vm.addQueueVersion = function() {
-        $rootScope.$emit('queue.query.reset');
         vm.selectedQueueVersion = vm.getDefaultVersion();
         vm.selectedQueueVersion.queueId = vm.selectedQueue.id;
         vm.selectedQueueVersion.name = 'v' + (vm.fetchVersions().length + 1);
       };
 
       vm.saveVersion = function() {
-        vm.selectedQueueVersion.query = ZermeloService.getQueryString();
         return vm.selectedQueueVersion.save().then(function(version) {
           vm.selectedQueueVersion = null;
-          $rootScope.$emit('queue.query.reset');
           return version;
         }).finally(vm.updateVersions);
       };